fix: layout shift caused by top banner
