+++ src/components/app/geoLocationDisclaimerBanner/index.tsx
@@ -1,0 +1,9 @@
+import { GeoLocationDisclaimerBannerContent } from '@/components/app/geoLocationDisclaimerBanner/geoLocationDisclaimerBannerContent'
+
+export function GeoLocationDisclaimerBanner() {
+  return (
+    <div className="flex h-12 w-full items-center justify-center bg-primary-cta">
+      <GeoLocationDisclaimerBannerContent />
+    </div>
+  )
+}
